+++ SeaHours/SeaHours/PlayerStatus.storyboard
@@ -7,10 +7,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--Game View Controller-->
+        <!--Status View Controller-->
         <scene sceneID="xTa-pI-Rct">
             <objects>
-                <viewController id="XvE-uk-eQd" customClass="GameViewController" customModule="SeaHours" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="XvE-uk-eQd" customClass="StatusViewController" customModule="SeaHours" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" multipleTouchEnabled="YES" contentMode="scaleToFill" id="16Q-H0-A4y" customClass="SKView" customModule="SeaHours">
                         <rect key="frame" x="0.0" y="0.0" width="1112" height="834"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
